mandatory part ended.

+++ ft_print_x.c
@@ -6,13 +6,13 @@
 /*   By: beldemir <beldemir@student.42istanbul.c    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/10/27 17:24:49 by beldemir          #+#    #+#             */
-/*   Updated: 2024/10/27 21:10:55 by beldemir         ###   ########.fr       */
+/*   Updated: 2024/10/28 02:24:06 by beldemir         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "ft_printf.h"
 
-int ft_print_x(int n, const char c)
+int	ft_print_x(unsigned long n, const char c)
 {
 	if (c == 'x')
 		return (ft_base(n, "0123456789abcdef"));
